pyproject.toml
[project]
name = "estate-pdf-organizer"
version = "0.1.0"
description = "A tool for processing and organizing estate documents using LLMs"
authors = [
    {name = "Jacob", email = "jacob@example.com"}
]
dependencies = [
    "click>=8.1.0",
    "pdfplumber>=0.10.0",
    "pypdf>=3.17.0",
    "langchain>=0.1.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.65.0",
    "loguru>=0.7.0",
    "pytest>=7.4.0",
]
requires-python = ">=3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"] 